Changed method to POST, transferring request params in body, all types changed to Double

+++ src/main/java/ua/bychkovskyy/math/Operation.java
@@ -1,5 +1,5 @@
 package ua.bychkovskyy.math;
 
 public interface Operation {
-    double doOperation(long a, long b);
+    double doOperation(double a, double b);
 }
