+++ src/main/java/ua/bychkovskyy/math/impl/AddOperation.java
@@ -8,7 +8,7 @@ public class AddOperation implements Operation {
     private final Logger logger = Logger.getLogger(getClass());
 
     @Override
-    public double doOperation(long a, long b) {
+    public double doOperation(double a, double b) {
         logger.info(String.format("Adding operands '%s' and '%s'", a, b));
         return a + b;
     }
