+++ src/main/java/ua/bychkovskyy/math/impl/MultiplicationOperation.java
@@ -8,7 +8,7 @@ public class MultiplicationOperation implements Operation {
     private final Logger logger = Logger.getLogger(getClass());
 
     @Override
-    public double doOperation(long a, long b) {
+    public double doOperation(double a, double b) {
         logger.info(String.format("Multiplication operands '%s' and '%s'", a, b));
         return a * b;
     }
